refactor: move code from calendarDatescontroller to the CalendarDateLogic file
