+++ backend/Controllers/CalendarDatesController.cs
@@ -1,5 +1,6 @@
 using backend.Data;
 using backend.Models;
+using backend.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -38,65 +39,21 @@ public class CalendarDatesController : ControllerBase
     [HttpGet("batch")]
     public ActionResult<CalendarDate?[]> GetCalendarDateBatch(DateTime start, DateTime end)
     {
-        //todo: make async?
-        var convertedDateStart = DateTime.SpecifyKind(start, DateTimeKind.Utc);
-        var convertedDateEnd = DateTime.SpecifyKind(end, DateTimeKind.Utc);
-
-        if (convertedDateEnd < convertedDateStart)
+        if (end < start)
         {
             return BadRequest("Start date has to be before end date");
         }
-
-        CalendarDate?[] dates = new CalendarDate[Convert.ToInt32(Math.Ceiling((convertedDateEnd - convertedDateStart).TotalDays)) + 1];
-
-        var dbDates = _context.CalendarDates
-                        .Include(convertedDate => convertedDate.DateContent)
-                        .ThenInclude(content => content.Events)
-                        .Where(calendarDate => calendarDate.Date.Date >= convertedDateStart.Date && calendarDate.Date.Date <= convertedDateEnd.Date).ToList();
-
-
-        for (int i = 0; i < dates.Length; i++)
-        {
-            var day = dbDates.FirstOrDefault(d => d.Date == convertedDateStart.AddDays(i));
-
-            if (day != null)
-            {
-                dates[i] = day;
-            }
-            else
-            {
-                dates[i] = null;
-            }
-        }
-        return dates;
-    }
-
-    private static DateTime GetMondayDate(int year, int weekNumber)
-    {
-        var cultureInfo = CultureInfo.InvariantCulture;
-        Calendar calendar = cultureInfo.Calendar;
-
-        CalendarWeekRule weekRule = CalendarWeekRule.FirstFourDayWeek;
-        DayOfWeek firstDayOfWeek = DayOfWeek.Monday;
-
-        DateTime jan4 = new DateTime(year, 1, 4);
-
-        int firstWeek = calendar.GetWeekOfYear(jan4, weekRule, firstDayOfWeek);
-        DateTime firstMonday = jan4.AddDays(-(int)(jan4.DayOfWeek - DayOfWeek.Monday));
-
-        DateTime mondayOfWeek = firstMonday.AddDays((weekNumber - firstWeek) * 7);
-
-        return mondayOfWeek;
+        return CalendarDateLogic.GetCalendarDateBatch(start, end, _context);
     }
 
     [HttpGet("Weeks/{weekNumber}")]
     public ActionResult<CalendarDate?[]> GetCalendarDate2Weeks(int weekNumber)
     {
-        var year = DateTime.Now.Year;
-        var mondayDate = GetMondayDate(year, weekNumber);
-        Console.WriteLine(mondayDate);
-
-        return GetCalendarDateBatch(mondayDate, mondayDate.AddDays(13));
+        if(weekNumber < 1 || weekNumber > 53)
+        {
+            return BadRequest("Weeknumber has to be between 1 and 53");
+        }
+        return CalendarDateLogic.GetCalendarDate2Weeks(weekNumber, _context);
     }
